fix(gb): dedupe S9xSGBGetSRAMSize after rebase
The libretro and debugger branches each added an S9xSGBGetSRAMSize;
the rebase merged both, producing a duplicate definition (C2084).
Drop the debugger copy, keep the canonical battery-API one, and cast
at the CDebuggerGb caller to avoid a size_t->uint32_t narrowing.

+++ win32/CDebuggerGb.cpp
@@ -38,7 +38,7 @@ uint32_t CurrentDisplayPC()
 
 uint32_t CartSRAMSize()
 {
-	return S9xSGBGetSRAMSize();
+	return (uint32_t)S9xSGBGetSRAMSize();
 }
 
 } // namespace GbBackend
